refactor: rename Footer => index.tsx
