+++ lib/components/Footer/index.tsx
